requirements.txt
ragchecker
openai==2.11.0
python-dotenv==1.2.1
httpx==0.28.1
pymilvus==2.6.5
transformers==4.57.3
sentence-transformers==5.2.0
datasets==4.4.1
tqdm==4.67.1

# PyTorch CPU
#--extra-index-url https://download.pytorch.org/whl/cpu
#torch
#torchvision

# PyTorch with CUDA 13.0
--extra-index-url https://download.pytorch.org/whl/cu130
torch
torchvision
xformers